Update Mentee Model
