+++ DTD_Mentorship_Project/DTD_Mentorship_Project/Models/Mentee.cs
@@ -1,0 +1,10 @@
+﻿namespace DTD_Mentorship_Project.Models {
+    public class Mentee
+{
+    public string? Name { get; set; }
+    public string? Description { get; set; }
+    public string? Title { get; set; }
+    public string? Area { get; set; }
+    public string? Image { get; set; }
+}
+}
