+++ DTD_Mentorship_Project/DTD_Mentorship_Project/Models/Mentor.cs
@@ -1,0 +1,6 @@
+﻿namespace DTD_Mentorship_Project.Models
+{
+    public class Mentor
+    {
+    }
+}
