+++ DTD_Mentorship_Project/DTD_Mentorship_Project/Pages/Dashboard.cshtml.cs
@@ -1,16 +1,10 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using DTD_Mentorship_Project.Models;
 
 namespace DTD_Mentorship_Project.Pages
 {
-    public class Mentee
-    {
-        public string? Name { get; set; }
-        public string? Description { get; set; }
-        public string? Title { get; set; }
-        public string? Area { get; set; }
-        public string? Image { get; set; }
-    }
+   
     public class DashboardModel : PageModel
     {
         private readonly ILogger<DashboardModel> _logger;
